Move dav utils functions to OCP\Files\DavUtil
