+++ lib/public/Util.php
@@ -628,53 +628,4 @@ class Util {
 		}
 		return true;
 	}
-
-	/**
-	 * Compute the fileId to use for dav responses
-	 *
-	 * @param int $id Id of the file returned by FileInfo::getId
-	 * @since 25.0.0
-	 */
-	public static function getDavFileId(int $id): string {
-		$instanceId = \OC_Util::getInstanceId();
-		$id = sprintf('%08d', $id);
-		return $id . $instanceId;
-	}
-
-	/**
-	 * Compute the format needed for returning permissions for dav
-	 *
-	 * @since 25.0.0
-	 */
-	public static function getDavPermissions(\OCP\Files\FileInfo $info): string {
-		$p = '';
-		if ($info->isShared()) {
-			$p .= 'S';
-		}
-		if ($info->isShareable()) {
-			$p .= 'R';
-		}
-		if ($info->isMounted()) {
-			$p .= 'M';
-		}
-		if ($info->isReadable()) {
-			$p .= 'G';
-		}
-		if ($info->isDeletable()) {
-			$p .= 'D';
-		}
-		if ($info->isUpdateable()) {
-			$p .= 'NV'; // Renameable, Moveable
-		}
-		if ($info->getType() === \OCP\Files\FileInfo::TYPE_FILE) {
-			if ($info->isUpdateable()) {
-				$p .= 'W';
-			}
-		} else {
-			if ($info->isCreatable()) {
-				$p .= 'CK';
-			}
-		}
-		return $p;
-	}
 }
